Banner Template

+++ app/Modules/Page/Service/PageService.php
@@ -5,6 +5,7 @@ namespace App\Modules\Page\Service;
 
 use App\Modules\Page\Entity\Page;
 use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 
 class PageService
 {
@@ -35,16 +36,55 @@ class PageService
         ]);
     }
 
-    public function destroy(Page $page)
+    public function destroy(Page $page): void
     {
+        if ($page->published) throw new \DomainException('страница опубликована');
         $page->delete();
     }
 
-    public function setText(Request $request, Page $page): Page
+    public function setText(Page $page, string $text): void
     {
-        $text = $request['text'];
-        $page->setText($text);
-        return $page;
+        $page->text = $text;
+        $page->save();
     }
 
+    public function setInfo(Page $page, Request $request): void
+    {
+        $name = $request->string('name')->trim()->value();
+        $slug = $request->string('slug')->trim()->value();
+        $page->name = $name;
+        $page->slug = empty($slug) ? Str::slug($name) : $slug;
+        $page->title = $request->string('title')->trim()->value();
+        $page->description = $request->string('description')->trim()->value();
+        $page->parent_id = $request->input('parent_id');
+        $page->menu = $request->boolean('menu');
+        $page->template = $request->string('template')->value();
+        $page->save();
+    }
+
+    public function up(Page $page): void
+    {
+        $pages = Page::orderBy('sort')->where('parent_id', $page->parent_id)->getModels();
+        for ($i = 1; $i < count($pages); $i++) {
+            if ($pages[$i]->id == $page->id) {
+                $prev = $pages[$i - 1]->sort;
+                $next = $pages[$i]->sort;
+                $pages[$i]->update(['sort' => $prev]);
+                $pages[$i - 1]->update(['sort' => $next]);
+            }
+        }
+    }
+
+    public function down(Page $page): void
+    {
+        $pages = Page::orderBy('sort')->where('parent_id', $page->parent_id)->getModels();
+        for ($i = 0; $i < count($pages) - 1; $i++) {
+            if ($pages[$i]->id == $page->id) {
+                $prev = $pages[$i + 1]->sort;
+                $next = $pages[$i]->sort;
+                $pages[$i]->update(['sort' => $prev]);
+                $pages[$i + 1]->update(['sort' => $next]);
+            }
+        }
+    }
 }
